+++ Runtime/RuntimeEvents.lua
@@ -2,7 +2,6 @@ local LifecycleEvents = ApogeePartyHealthBars_LifecycleEvents
 local UnitEvents = ApogeePartyHealthBars_UnitEvents
 local ActionEvents = ApogeePartyHealthBars_ActionEvents
 local ActionAssignmentEvents = ApogeePartyHealthBars_ActionAssignmentEvents
-local DungeonBoardEvents = ApogeePartyHealthBars_DungeonBoardEvents
 local CleanseEvents = ApogeePartyHealthBars_CleanseEvents
 local GroupHelperEvents = ApogeePartyHealthBars.Require("Runtime", "GroupHelperEvents")
 
@@ -20,7 +19,6 @@ function R.Register(eventRouter, deps)
         RefreshAssignmentAffordances = deps.RefreshAssignmentAffordances,
     })
     ActionEvents.Register(eventRouter, deps)
-    if DungeonBoardEvents then DungeonBoardEvents.Register(eventRouter, deps) end
     if CleanseEvents then CleanseEvents.Register(eventRouter, deps) end
     GroupHelperEvents.Register(eventRouter, deps)
 end